Create fbref_web_scrape.py
This code scrapes fixture result data from FBref and outputs a csv file. However, it is not suitable for leagues with play-offs (Knock-Out matches). It can be used for leagues in every country and season available in FBref.

Knockout matches can also be included with slight modifications.

PS: There is lots of room for further optimizations